rename PRMSSolarGeometry and PRMSAtmosphere modules
